Add files via upload
Program that determnines if you've lived long enough in a city and that you're old enough to be a part of the city council or run for mayor. You need to be atleast 30 years old and lived in the city atleast 9 years to become Mayor. To become a part of the City Council you'll need to be atleast 25 years old and lived in the city for atleast 5 years.